Move funcs from child/console -> utils
