+++ packages/common/utils.js
@@ -18,3 +18,24 @@ export const round = (value) => Math.round(value * ROUNDING) / ROUNDING
 
 export const capitalizeFirstLetter = (string) =>
   string.charAt(0).toUpperCase() + string.slice(1)
+
+const isDef = (value) => `${value}` !== '' && value !== undefined
+
+export function getElementName(el) {
+  switch (true) {
+    case !isDef(el):
+      return ''
+
+    case isDef(el.id):
+      return `${el.nodeName.toUpperCase()}#${el.id}`
+
+    case isDef(el.name):
+      return `${el.nodeName.toUpperCase()} (${el.name})`
+
+    default:
+      return (
+        el.nodeName.toUpperCase() +
+        (isDef(el.className) ? `.${el.className}` : '')
+      )
+  }
+}
